requirements.txt
cryptography
pillow
matplotlib
numpy